Add instructions on local deployment similar to Appengine
Add instructions to README.md on how to run a local deployment similar to
Appengine envieonment because it was useful to test redirects.
Update "running locally" section for deployment outside out of Appengine so
instructions are consistent with Appengine instructions. `go run ...` appears
to be as performant but shorter compared to `go build ...` followed by
`./go-slides ...`.